aws_update_route53/requirements.txt
boto
dnspython
netifaces
